refactor: extract version_cmp and mpsc_invariants from connection_handler.rs
connection_handler.rs was 5747 lines. Two isolated blocks had no
dependencies on the rest of the file and could be extracted verbatim:

- mod version_cmp (~963 lines): wire-format version encoding/decoding,
  depends only on config constants. Moved to
  crates/core/src/transport/connection_handler/version_cmp.rs.

- mod mpsc_invariants (~359 lines): #[cfg(test)] smoke tests pinning
  tokio::sync::mpsc invariants, introduced in #3960. Moved to
  crates/core/src/transport/connection_handler/tests/mpsc_invariants.rs
  under a new connection_handler/tests.rs module gateway.

Uses foo.rs + foo/ layout as required by code-style.md (no mod.rs).
Content moved byte-for-byte, dedented one level. No logic changed.
mock_transport and the listener/handshake state-machine deliberately
left in place (Phase 2 scope).

Pre-commit clippy hook fails on pre-existing clippy-1.94 lints in
service.rs, update.rs, user_op_rate_limit.rs, and module_cache.rs
(verified on main/HEAD: same 7 errors, none in transport/). The hook
runs --all-targets --features bench which hits test-only code already
broken on main. cargo clippy -p freenet --lib passes clean.
